auth fix

+++ src/app/api/(files)/lessons/[id]/route.ts
@@ -1,5 +1,5 @@
-import { auth } from "@clerk/nextjs";
-import { Lesson } from "@prisma/client";
+import { auth, currentUser } from "@clerk/nextjs";
+import { Lesson, Role } from "@prisma/client";
 import { revalidateTag } from "next/cache";
 import { NextResponse } from "next/server";
 import prisma from "src/utils/prisma";
@@ -23,7 +23,15 @@ export async function PATCH(
   request: Request,
   { params }: { params: { id: string } }
 ) {
-  const { userId } = auth();
+  const user = await currentUser();
+  if (!user?.id)
+    return NextResponse.json({ message: "user not found" }, { status: 400 });
+  const role = user?.publicMetadata?.role as Role;
+  if (role !== Role.ADMIN)
+    return NextResponse.json(
+      { message: "Only admins have permission to update" },
+      { status: 403 }
+    );
   const res = (await request.json()) as Partial<Lesson>;
   const updateData = Object.fromEntries(
     Object.entries(res).map(([key, value]) => [key, { set: value }])
@@ -36,7 +44,6 @@ export async function PATCH(
     },
     data: updateData,
   });
-
 
   if (data.subtopicId) {
     revalidateTag(`subtopics/${data.subtopicId}`);
@@ -54,6 +61,15 @@ export async function DELETE(
   request: Request,
   { params }: { params: { id: string } }
 ) {
+  const user = await currentUser();
+  if (!user?.id)
+    return NextResponse.json({ message: "user not found" }, { status: 400 });
+  const role = user?.publicMetadata?.role as Role;
+  if (role !== Role.ADMIN)
+    return NextResponse.json(
+      { message: "Only admins have permission to update" },
+      { status: 403 }
+    );
   const id = Number(params.id);
   const data = await prisma.lesson.delete({
     where: {
